+++ resources/views/bootstrap-4/form-select.blade.php
@@ -4,12 +4,16 @@
     <select
         @if($isWired())
             wire:model{!! $wireModifier() !!}="{{ $name }}"
-        @else
-            name="{{ $name }}"
         @endif
+
+        name="{{ $name }}"
 
         @if($multiple)
             multiple
+        @endif
+
+        @if($placeholder)
+            placeholder="{{ $placeholder }}"
         @endif
 
         @if($label && !$attributes->get('id'))
@@ -17,6 +21,13 @@
         @endif
 
         {!! $attributes->merge(['class' => 'form-control ' . ($hasError($name) ? 'is-invalid' : '')]) !!}>
+
+        @if($placeholder)
+            <option value="" disabled @if($nothingSelected()) selected="selected" @endif>
+                {{ $placeholder }}
+            </option>
+        @endif
+
         @forelse($options as $key => $option)
             <option value="{{ $key }}" @if($isSelected($key)) selected="selected" @endif>
                 {{ $option }}