+++ src/Components/FormRadio.php
@@ -27,12 +27,12 @@ class FormRadio extends Component
 
         $inputName = static::convertBracketsToDots($name);
 
-        if (old($inputName)) {
+        if (old($inputName) !== null) {
             $this->checked = old($inputName) == $value;
         }
 
         if (!session()->hasOldInput() && $this->isNotWired()) {
-            $boundValue = $this->getBoundValue($bind, $name);
+            $boundValue = $this->getBoundValue($bind, $inputName);
 
             if (!is_null($boundValue)) {
                 $this->checked = $boundValue == $this->value;